+++ src/routes/index.ts
@@ -1,0 +1,3 @@
+export * from './routes-config';
+export * from './error-page';
+export * from './root-layout';
